requirements.txt
Flask==3.0.2
qrcode[pil]==7.4.2
python-dotenv==1.0.1